Extract session task state store
